requirements.txt
torch==1.8.1+cu111
torchvision==0.9.1+cu111
numpy
albumentations
pytorch_lightning
wandb
pandas
opencv_python
tokenizers
python_Levenshtein
PyYAML
